Significantly enhance scorecard generator:   1) Add commercial average score calculation   2) Add ability to create anonymous scorecards for commercial tools   3) Add ability to suppress commercial tool scorecards and only include the commercial average   4) Add details table to Home and Vulnerability pages

+++ src/main/java/org/owasp/benchmark/score/parsers/Counter.java
@@ -20,9 +20,9 @@ package org.owasp.benchmark.score.parsers;
 
 public class Counter {
 
-	public double tp = 0;
-	public double fn = 0;
-	public double tn = 0;
-	public double fp = 0;
+	public int tp = 0;
+	public int fn = 0;
+	public int tn = 0;
+	public int fp = 0;
 	
 }
